Convert to double

+++ definitions.cpp
@@ -9,11 +9,11 @@ using namespace std;
 
 
 void folds_stats::logavg() {
-    // for random num generating. Only needs to be called once.
+    // Seed the RNG
     srand(static_cast <unsigned> (time(0)));
 
     // fold count is just a vector of integers increasing by 1.
-    for (float i = 0; i < n; i++) {
+    for (int i = 0; i < n; i++) {
         f.insert(f.end(), {i + 1});
     }
 
@@ -23,18 +23,18 @@ void folds_stats::logavg() {
         logc.insert(logc.end(), { 0 });
     } 
 
-    // loop (instance) times to average log(crease values) for accurate statistics. 
+    // Loop (instance) times to average log(crease values) for accurate statistics. 
     for (int j = 1; j <= instance; j++) {
         segs_i = { 0,1 };
         // Fold n times. Count creases and take their logs
         for (int i = 0; i < n; i++) {
             segs_o = fold(segs_i);
             sizeo = segs_o.size();
-            // number of creases is number of segments - 1
+            // Number of creases is number of segments - 1
             c[i] = (sizeo / 2) - 1;
             logc[i] = log(c[i]);
             cavg[i] += logc[i];
-            // avoid the case when no fold happened the first time. then log(0) = -inf
+            // Avoid the case when no fold happened the first time. Then log(0) = -inf
             if (c[i] == 0.0) {
                 c[i] = 1;
                 logc[i] = 0;
@@ -53,17 +53,16 @@ void folds_stats::logavg() {
     }
     data.close();
 }
-
-// fold function.  
-vector<float> folds_stats::fold(vector<float> &segs_in) {
-    vector<float> segs_out;
-    // 1 for fold direction left; 0 for right. rand() is seeded in main()
+ 
+vector<double> folds_stats::fold(vector<double> &segs_in) {
+    vector<double> segs_out;
+    // 1 for fold direction left; 0 for right.
     direct = rand() % 2;
    
     // generates random fold pos.
     LO = getmin(segs_in);
     HI = getmax(segs_in);
-    x = LO + static_cast <float> (rand()) / (static_cast <float> (RAND_MAX / (HI - LO)));
+    x = LO + static_cast <double> (rand()) / (static_cast <double> (RAND_MAX / (HI - LO)));
 
     // check if x in outermost segment
     if (LO < x && x < HI) {
@@ -118,7 +117,7 @@ vector<float> folds_stats::fold(vector<float> &segs_in) {
 }
 
 // Find min element at beginning of fold() 
-float folds_stats::getmin(vector<float> &segs_in) {
+double folds_stats::getmin(vector<double> &segs_in) {
     size = segs_in.size();
     min = segs_in[0];
     for (int i = 1; i < size; i++) {
@@ -130,7 +129,7 @@ float folds_stats::getmin(vector<float> &segs_in) {
 }
 
 // Find max element at beginning of fold()
-float folds_stats::getmax(vector<float> &segs_in) {
+double folds_stats::getmax(vector<double> &segs_in) {
     size = segs_in.size();
     max = segs_in[0];
     for (int i = 1; i < size; i++) {
@@ -142,8 +141,8 @@ float folds_stats::getmax(vector<float> &segs_in) {
 }
 
 
-// display vec function. For debug purposes only
-void folds_stats::display(vector<float> arr) {
+// Display vec function. For debug purposes only
+void folds_stats::display(vector<double> arr) {
     for (int i = 0; i < arr.size(); i++) {
         cout << arr[i] << '\n';
     }
